+++ CHANGES.md
@@ -10,6 +10,7 @@
   - reduce signature of `OAuthClient` constructor, introduce `setSession`, 
     `setRandom` and `setDateTime`. Note that these are only for testing or 
    very special use cases, so most likely not for normal API users!
+- remove `OAuthClient::hasAccessToken` as no code actually uses it
 
 ## 6.0.2 (2018-03-21)
 - use safe `strlen` from `paragonie/constant_time_encoding`